Move collapse thinking setting to appearance messages section
